+++ lib/bin/stats.ts
@@ -1,5 +1,9 @@
 import buildStats from '../stats'
 
-const stats = () => buildStats()
+const stats = (...params: string[]) => {
+    const verbose = params.indexOf('verbose') !== -1
+
+    return buildStats(verbose)
+}
 
 export default stats