+++ lib/stats/ssr-stats.ts
@@ -9,7 +9,7 @@ import { runStatsOnServer, loadSSRPage } from './server'
 
 const { HAS_SERVER } = CONFIG
 
-const ssrStats = async (): Promise<BuildMetrics> => {
+const ssrStats = async (verbose = false): Promise<BuildMetrics> => {
 
     if (!HAS_SERVER) {
         log('Skipping SSR stats because the application has no server')
@@ -27,6 +27,13 @@ const ssrStats = async (): Promise<BuildMetrics> => {
 
             const { size, content } = await loadPage()
 
+            if (verbose) {
+                log('### SSR content for ', url)
+                log('---------------------')
+                log(content)
+                log('---------------------')
+            }
+
             const domSize = load(content)('*').length
 
             const time = await getSequenceAverage(async () => {
@@ -37,7 +44,7 @@ const ssrStats = async (): Promise<BuildMetrics> => {
             stats[`${page}_ssr_document_size`] = formatFileSize(size)
             stats[`${page}_ssr_dom_size`] = domSize.toString()
             stats[`${page}_ssr_loadtime`] = formatTimeMs(time)
-        })
+        }, verbose)
 
         log(`SSR stats: ${prettyJson(stats)}`)